[daemons] E: Send Interrupt Signal To Linuxd Worker Threads
In this commit I implement support to interrupting worker threads in
linuxd. This change involves two moving pieces:
1) Installing a SIGUSR handler in worker threads, as well as a mechanism
   to kill them (using `pthread_kill`).
2) Handling EINTR failures in all blocking system calls accordingly.

To accomodate for 1) I move the per-worker-thread logic from
`linuxd/mod.rs` to a dedicated `worker_thread.rs` file.

To properly propagate EINTR exits up the stack, I modify all system call
implementations to return a Result<Message, WorkerThreadError> that can
be used across the codebase.

This is a first step towards properly cleaning-up resources in linuxd
when a user VM shuts down.
